# tacocloud/tacocloud-api/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Taco Cloud HTTP API
  version: 1.0.0
  description: |
    Contrato público de /api/v1. Las rutas /api equivalentes siguen disponibles
    durante la transición y responden con Deprecation y Link al sucesor.
    Las escrituras autenticadas por sesión requieren X-XSRF-TOKEN. El servidor
    acepta X-Correlation-Id y devuelve el identificador efectivo en ese header.
    GET /api/v1/tacos?recent conserva la respuesta de lista del cliente antiguo.
servers:
  - url: /
security:
  - basicAuth: []
...
paths:
...
  /api/v1/tacos:
    get:
      tags: [Tacos]
      summary: Buscar tacos publicados
      description: Con el parámetro recent se devuelve una lista de hasta 12 tacos.
      security: []
      parameters:
        - { name: name, in: query, schema: { type: string, maxLength: 100 } }
        - { name: ingredientId, in: query, schema: { type: string } }
        - { name: diet, in: query, schema: { $ref: '#/components/schemas/DietaryTag' } }
        - { name: excludeAllergen, in: query, schema: { $ref: '#/components/schemas/Allergen' } }
        - { name: spice, in: query, schema: { $ref: '#/components/schemas/SpiceLevel' } }
        - { name: page, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, default: 20 } }
        - { name: sort, in: query, schema: { type: string, default: 'createdAt,desc' } }
        - { name: recent, in: query, schema: { type: string }, description: Modo legado de lista. }
      responses:
        '200':
          description: Página de búsqueda o lista recent.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TacoSearchResponse'
                  - type: array
                    items: { $ref: '#/components/schemas/TacoResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
    post:
      tags: [Tacos]
      summary: Publicar un taco en el catálogo
      x-required-role: ADMIN
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody: { $ref: '#/components/requestBodies/TacoCreateRequest' }
      responses:
        '201': { $ref: '#/components/responses/TacoCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /api/v1/tacos/validate:
    post:
      tags: [Tacos]
      summary: Validar un diseño sin guardarlo
      x-required-role: USER o ADMIN
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody: { $ref: '#/components/requestBodies/TacoCreateRequest' }
      responses:
        '200': { $ref: '#/components/responses/TacoDesignValidation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/tacos/top:
    get:
      tags: [Tacos, Ratings]
      summary: Consultar los tacos mejor valorados
      security: []
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, default: 10 } }
      responses:
        '200': { $ref: '#/components/responses/TacoRankingList' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/tacos/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Tacos]
      summary: Consultar un taco publicado
      security: []
      responses:
        '200': { $ref: '#/components/responses/Taco' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/tacos/{id}/classification:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Tacos]
      summary: Consultar clasificación dietaria y alérgenos
      security: []
      responses:
        '200': { $ref: '#/components/responses/TacoClassification' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/tacos/{id}/rating:
    parameters:
      - $ref: '#/components/parameters/Id'
    put:
      tags: [Ratings]
      summary: Valorar un taco de 1 a 5
      x-required-role: USER
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody: { $ref: '#/components/requestBodies/TacoRatingRequest' }
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/orders/quote:
    post:
      tags: [Orders]
      summary: Cotizar cupón y clasificación de taco
      x-required-role: USER
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody: { $ref: '#/components/requestBodies/OrderQuoteRequest' }
      responses:
        '200': { $ref: '#/components/responses/OrderQuote' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unprocessable' }
...
  /api/v1/orders/me/{sourceOrderId}/reorder/quote:
    parameters:
      - { name: sourceOrderId, in: path, required: true, schema: { type: string } }
    post:
      tags: [Orders]
      summary: Cotizar una recompra
      x-required-role: USER
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody: { $ref: '#/components/requestBodies/ReorderRequest' }
      responses:
        '200': { $ref: '#/components/responses/ReorderQuote' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/orders/me/{sourceOrderId}/reorder:
    parameters:
      - { name: sourceOrderId, in: path, required: true, schema: { type: string } }
    post:
      tags: [Orders]
      summary: Confirmar una recompra idempotente
      x-required-role: USER
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody: { $ref: '#/components/requestBodies/ReorderRequest' }
      responses:
        '201': { $ref: '#/components/responses/ReorderConfirm' }
        '200': { $ref: '#/components/responses/ReorderConfirm' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Autenticación Basic o sesión autenticada de la UI.
  headers:
    CorrelationId:
      description: Identificador de correlación efectivo para esta petición.
      schema: { type: string, maxLength: 100 }
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema: { type: string, pattern: '^[A-Za-z0-9._:-]{1,100}$' }
      description: El servidor genera un ID si el recibido no es válido.
...
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128, pattern: '^[A-Za-z0-9_-]+$' }
    CsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: true
      schema: { type: string }
      description: Requerido para escrituras autenticadas por sesión. Se obtiene de la cookie XSRF-TOKEN.
  requestBodies:
...
    TacoCreateRequest:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/TacoCreateRequest' }
          example: { name: Garden taco, ingredients: [{ id: WRAP }, { id: SLSA }] }
    TacoRatingRequest:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/TacoRatingRequest' }
          example: { score: 5 }
...
    OrderQuoteRequest:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/OrderQuoteRequest' }
          example: { subtotal: 12.50, couponCode: CLASS10, tacoId: TACO-ID }
...
    ReorderRequest:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ReorderRequest' }
...
  responses:
    NoContent:
      description: Operación completada sin cuerpo.
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
...
    Taco:
      description: Taco publicado.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/TacoResponse' }
    TacoCreated:
      description: Taco publicado.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/TacoResponse' }
    TacoDesignValidation:
      description: Violaciones del diseño, si existen.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/TacoDesignValidationResponse' }
...
    TacoClassification:
      description: Clasificación de un taco.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/TacoClassification' }
    TacoRankingList:
      description: Tacos mejor valorados.
      content:
        application/json:
          schema:
            type: array
            items: { $ref: '#/components/schemas/TacoRankingResponse' }
...
    OrderQuote:
      description: Cotización sin persistir la orden.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/OrderQuoteResponse' }
...
    ReorderQuote:
      description: Cotización de recompra.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ReorderQuoteResponse' }
    ReorderConfirm:
      description: Orden nueva o repetición idempotente.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ReorderConfirmResponse' }
...
    BadRequest:
      description: Entrada inválida.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiProblem' }
          example: { type: 'urn:tacocloud:problem:malformed-request', title: Bad request, status: 400, detail: The request cannot be processed., instance: /api/v1/orders, code: MALFORMED_REQUEST, violations: [] }
...
    Forbidden:
      description: Operación prohibida. Spring Security puede devolver un 403 sin ApiProblem.
    NotFound:
      description: Recurso inexistente.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiProblem' }
    Conflict:
      description: Conflicto de estado, versión o Idempotency-Key.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiProblem' }
    Unprocessable:
      description: Regla de negocio o diseño inválido.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiProblem' }
  schemas:
    IngredientType:
      type: string
      enum: [WRAP, BOWL, PROTEIN, VEGGIES, CHEESE, SAUCE]
    DietaryTag:
      type: string
      enum: [VEGAN, VEGETARIAN, GLUTEN_FREE]
    Allergen:
      type: string
      enum: [GLUTEN, MILK, EGG, SOY, PEANUTS, TREE_NUTS, FISH, SHELLFISH, SESAME]
    SpiceLevel:
      type: string
      enum: [UNKNOWN, NONE, MILD, MEDIUM, HOT, EXTREME]
    OrderStatus:
      type: string
      enum: [CREATED, ACCEPTED, PLACED, PREPARING, READY, OUT_FOR_DELIVERY, DELIVERED, CANCELLED]
...
    IngredientResponse:
      type: object
      additionalProperties: false
      properties:
        id: { type: string }
        name: { type: string }
        type: { $ref: '#/components/schemas/IngredientType' }
        unitPrice: { type: number, format: decimal }
        available: { type: boolean }
        dietaryTags:
          type: array
          uniqueItems: true
          items: { $ref: '#/components/schemas/DietaryTag' }
        allergens:
          type: array
          uniqueItems: true
          items: { $ref: '#/components/schemas/Allergen' }
        spiceLevel: { $ref: '#/components/schemas/SpiceLevel' }
...
    TacoIngredientId:
      type: object
      additionalProperties: false
      required: [id]
      properties:
        id: { type: string, maxLength: 64, pattern: '^[A-Za-z0-9_-]+$' }
    TacoCreateRequest:
      type: object
      additionalProperties: false
      required: [name, ingredients]
      properties:
        name: { type: string, minLength: 5, maxLength: 100 }
        ingredients:
          type: array
          maxItems: 50
          items: { $ref: '#/components/schemas/TacoIngredientId' }
    TacoClassification:
      type: object
      additionalProperties: false
      properties:
        tacoId: { type: string, nullable: true }
        dietaryTags:
          type: array
          uniqueItems: true
          items: { $ref: '#/components/schemas/DietaryTag' }
        allergens:
          type: array
          uniqueItems: true
          items: { $ref: '#/components/schemas/Allergen' }
        spiceLevel: { $ref: '#/components/schemas/SpiceLevel' }
        disclaimer: { type: string }
    TacoResponse:
      type: object
      additionalProperties: false
      properties:
        id: { type: string }
        name: { type: string }
        createdAt: { type: string, format: date-time }
        ingredients:
          type: array
          items: { $ref: '#/components/schemas/IngredientResponse' }
        classification: { $ref: '#/components/schemas/TacoClassification' }
    TacoSearchResponse:
      type: object
      additionalProperties: false
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/TacoResponse' }
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
    TacoDesignViolation:
      type: object
      additionalProperties: false
      properties:
        code: { type: string }
        message: { type: string }
    TacoDesignValidationResponse:
      type: object
      additionalProperties: false
      properties:
        valid: { type: boolean }
        violations:
          type: array
          items: { $ref: '#/components/schemas/TacoDesignViolation' }
...
    TacoRatingRequest:
      type: object
      additionalProperties: false
      required: [score]
      properties:
        score: { type: integer, minimum: 1, maximum: 5 }
    TacoRankingResponse:
      type: object
      additionalProperties: false
      properties:
        tacoId: { type: string }
        tacoName: { type: string }
        average: { type: number, format: decimal }
        count: { type: integer, format: int64 }
        distribution:
          type: array
          items: { type: integer, format: int64 }
...
    OrderResponseTaco:
      type: object
      additionalProperties: false
      properties:
        id: { type: string, nullable: true }
        name: { type: string }
        ingredientIds:
          type: array
          items: { type: string }
    OrderResponseItem:
      type: object
      additionalProperties: false
      properties:
        taco: { $ref: '#/components/schemas/OrderResponseTaco' }
        quantity: { type: integer, minimum: 1 }
        unitPriceAtPurchase: { type: number, format: decimal }
        subtotal: { type: number, format: decimal }
    OrderStatusChange:
      type: object
      additionalProperties: false
      properties:
        from: { $ref: '#/components/schemas/OrderStatus' }
        to: { $ref: '#/components/schemas/OrderStatus' }
        actorId: { type: string }
        actorRole: { type: string }
        changedAt: { type: string, format: date-time }
        origin: { type: string }
        reason: { type: string }
    OrderResponse:
      type: object
      additionalProperties: false
      properties:
        id: { type: string }
        placedAt: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/OrderStatus' }
        version: { type: integer, format: int64 }
        statusHistory:
          type: array
          items: { $ref: '#/components/schemas/OrderStatusChange' }
        userId: { type: string }
        deliveryName: { type: string }
        deliveryStreet: { type: string }
        deliveryCity: { type: string }
        deliveryState: { type: string }
        deliveryZip: { type: string }
        paymentBrand: { type: string }
        paymentLast4: { type: string }
        currency: { type: string }
        subtotal: { type: number, format: decimal }
        couponApplied: { type: boolean }
        discount: { type: number, format: decimal }
        total: { type: number, format: decimal }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderResponseItem' }
...
    OrderQuoteRequest:
      type: object
      additionalProperties: false
      required: [subtotal, couponCode]
      properties:
        subtotal: { type: number, format: decimal, minimum: 0 }
        couponCode: { type: string, maxLength: 64, pattern: '^[A-Za-z0-9_-]+$' }
        tacoId: { type: string, maxLength: 64, pattern: '^[A-Za-z0-9_-]+$' }
        taco: { $ref: '#/components/schemas/TacoCreateRequest' }
    OrderQuoteResponse:
      type: object
      additionalProperties: false
      properties:
        currency: { type: string }
        subtotal: { type: number, format: decimal }
        discount: { type: number, format: decimal }
        total: { type: number, format: decimal }
        couponApplied: { type: boolean }
        classification: { $ref: '#/components/schemas/TacoClassification' }
...
    ReorderRequest:
      type: object
      additionalProperties: false
      required: [paymentMethodId]
      properties:
        paymentMethodId: { type: string, maxLength: 64 }
        couponCode: { type: string, maxLength: 64, pattern: '^[A-Za-z0-9_-]+$' }
        quoteFingerprint: { type: string, pattern: '^[0-9a-f]{64}$' }
    IngredientDifference:
      type: object
      additionalProperties: false
      properties:
        ingredientId: { type: string }
        previousName: { type: string }
        currentName: { type: string }
        previousUnitPrice: { type: number, format: decimal }
        currentUnitPrice: { type: number, format: decimal }
        metadataChanged: { type: boolean }
    ReorderQuoteResponse:
      type: object
      additionalProperties: false
      properties:
        sourceOrderId: { type: string }
        currency: { type: string }
        originalSubtotal: { type: number, format: decimal }
        quotedSubtotal: { type: number, format: decimal }
        originalTotal: { type: number, format: decimal }
        quotedTotal: { type: number, format: decimal }
        priceDifference: { type: number, format: decimal }
        couponApplied: { type: boolean }
        ingredientDifferences:
          type: array
          items: { $ref: '#/components/schemas/IngredientDifference' }
        quoteFingerprint: { type: string }
        confirmationRequired: { type: boolean }
    ReorderConfirmResponse:
      type: object
      additionalProperties: false
      properties:
        order: { $ref: '#/components/schemas/OrderResponse' }
        replayed: { type: boolean }
...
    ApiProblemViolation:
      type: object
      additionalProperties: false
      properties:
        field: { type: string }
        message: { type: string }
        code: { type: string, nullable: true }
    ApiProblem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, violations]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        violations:
          type: array
          items: { $ref: '#/components/schemas/ApiProblemViolation' }